Bot/requirements.txt
python-telegram-bot>=20.0
yt-dlp
redis
python-dotenv